create some pages and create dynamic routes for classRoom page
