+++ .prettierrc
@@ -1,1 +1,1 @@
-{ "printWidth": 100 }
+{ "printWidth": 120 }
