+++ src/pages/cart.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function cart() {
+  return <div>cart</div>;
+}
+
+export default cart;
