--- src/pages/myCourses.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function myCourses() {
-  return <div>myCourse</div>;
-}
-
-export default myCourses;
