+++ src/pages/reserve.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function reserve() {
+  return <div>reserve</div>;
+}
+
+export default reserve;
