+++ src/pages/videoCourses.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function videoCourses() {
+  return <div>videoCourses</div>;
+}
+
+export default videoCourses;
